+++ src/Term.hs
@@ -650,6 +650,7 @@ branch = do
 
 parseTerm = parse term "Parse error"
 
+parseModule :: String -> Either ParseError ([[Char]], [(String, ([String], Term))])
 parseModule = parse modul "Parse error"
