+++ test/DistillEvalEqTest.hs
@@ -1,9 +1,5 @@
-module EvalTest where
+module DistillEvalEqTest where
 
 import Test.HUnit
 import Term
 
-
-distillEvalEqTest1 :: Test
-distillEvalEqTest1 = parsemodule
--- integration tests
