+++ test/Spec.hs
@@ -1,2 +1,10 @@
+import EvalTest
+
+import Test.HUnit
+import Test.Framework
+import Test.Framework.Providers.HUnit
+
 main :: IO ()
-main = print "Hello!"
+main = do
+    tests <- EvalTest.eval1Test 
+    defaultMainWithOpts (hUnitTestToTests tests) mempty
